Fix type conversion errors and remove unused code
- Fix int/double type conversion errors in salary_entitlement_service.dart
- Fix num to int/double conversions in sync_service.dart
- Fix double parameter type errors in enhanced_pdf_helper.dart
- Remove unused import and private method tests in booking_price_adjustment_test.dart
- Remove unused _calculateNightlyRateForSegment method in restore_fix_service.dart

+++ mobile/lib/services/salary_entitlement_service.dart
@@ -70,7 +70,7 @@ class SalaryEntitlementService {
     }
 
     final totalMonthsWorked = _calculateMonthsDifference(hireDate, now);
-    final totalEntitlement = totalMonthsWorked * employee.basicSalary;
+    final totalEntitlement = (totalMonthsWorked * employee.basicSalary).toDouble();
 
     final expenses = await (_db.select(
       _db.expenses,
